Fix form validation
Having two redux form instances mounted with the same name, but with
separate options was messing with the validations.  After each change,
the initial `reduxForm` defined in `CheckoutForm` would run its
validations setting the correct errors. Immediately afterwards, the
`reduxForm` defined in `CardElement` would unset the validations as it
had no validation function defined.

The selected solution is to have a single `reduxForm` defined (in
`CheckoutForm`), and pass it as a prop to `CardElement`.