Don't hardcode constants at build time
When we were an app, it was OK to hardcode certain constants at build
time.  Each deployment, staging, production, etc., would be configured
with environment variables.  When a new version was deployed, the app
would be built and those environment variables would be hardcoded into
the app.  This meant that each deployment had its own build of the app.

We now want a *single* build that can be added to Flight Center and used
correctly in FCs development, staging and production environments
without further modification.  This requires the configuration, to be
passed to the library at runtime.

This commit adds a new module `constants` which deals with getting and
setting "constants".  They are constant in the sense that 1) they used
to be; 2) its expected that they don't change during the runtime of the
component.