Prevent demo configuration from leaking
The demo application has some associated configuration which we don't
want to leak into the library.  In particular, the configuration of
the SSO cookie name.

Previously, we were using `dotenv` in a manner similar to
`create-react-app`.  Now we hardcode the configuration in the demo app
itself.  This prevents the demo configuration from leaking and makes the
demo easier to understand.